Project convert from TypeScript To Javascript with (with help of @augmentcode)
