+++ .editorconfig
@@ -1,6 +1,6 @@
 root = true
 
-[*.{js, ts}]
+[*.js]
 charset = utf-8
 indent_style = space
 indent_size = 2
