+++ .github/workflows/on-pull-request.yml
@@ -4,7 +4,7 @@ on:
   pull_request:
     branches:
       - main
-      
+
 permissions: write-all
 
 jobs:
@@ -13,14 +13,12 @@ jobs:
     strategy:
       matrix:
         node-version:
-          - 16.x
-          - 18.x
-          - 19.x
+          - 20.x
     steps:
       - name: Checkouting code...
         uses: actions/checkout@v4
         with:
-          ref: ${{ github.head_ref }}
+          ref: ${{ github.ref }}
 
       - name: Use Node.js ${{ matrix.node-version }}
         uses: actions/setup-node@v4
@@ -40,14 +38,14 @@ jobs:
 
     strategy:
       matrix:
-        node-version: 
-          - 16.x
+        node-version:
+          - 20.x
 
     steps:
       - name: Checkouting code...
         uses: actions/checkout@v4
         with:
-          ref: ${{ github.head_ref }}
+          ref: ${{ github.ref }}
 
       - name: Use Node.js ${{ matrix.node-version }}
         uses: actions/setup-node@v4
@@ -56,9 +54,6 @@ jobs:
 
       - name: NPM install
         run: npm install
-
-      - name: NPM build
-        run: npm run build
 
       - name: NPM format
         run: npm run format
@@ -88,25 +83,25 @@ jobs:
       - name: Checkouting code...
         uses: actions/checkout@v4
         with:
-          ref: ${{ github.head_ref }}
+          ref: ${{ github.ref }}
 
       - name: Downloading build artifact....
         uses: actions/download-artifact@v4
         with:
           name: build
           path: dist/
-          
+
       - name: Auto commiting changes...
         uses: stefanzweifel/git-auto-commit-action@v5
         with:
           commit_message: Automatically builded and updated
-          file_pattern: dist/*.js
+          file_pattern: dist/*.*
           skip_fetch: true
           skip_checkout: true
 
   dependabot:
-    needs: 
-      - build    
+    needs:
+      - build
       - commit_and_push
     permissions:
       pull-requests: write
@@ -136,4 +131,4 @@ jobs:
         run: gh pr merge --auto --squash "$PR_URL"
         env:
           PR_URL: ${{ github.event.pull_request.html_url }}
-          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}    
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
