+++ .prettierrc.json
@@ -7,5 +7,5 @@
   "trailingComma": "none",
   "bracketSpacing": false,
   "arrowParens": "avoid",
-  "parser": "typescript"
+  "parser": "babel"
 }
