+++ action.yml
@@ -10,5 +10,5 @@ inputs:
     required: true
 
 runs:
-  using: 'node16'
+  using: 'node20'
   main: 'dist/index.js'
