--- src/types/recursive-readdir-sync.d.ts
@@ -1,5 +1,0 @@
-declare module "recursive-readdir-sync" {
-
-  export default function (filesPath: string): string[]
-
-}
