--- tsconfig.build.json
@@ -1,6 +1,0 @@
-{
-  "extends": "./tsconfig.json",  
-  "exclude": [
-    "tests"
-  ]
-}
