--- tsconfig.json
@@ -1,22 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "ES2018",
-    /* Specify ECMAScript target version: 'ES3' (default), 'ES5', 'ES2015', 'ES2016', 'ES2017', 'ES2018', 'ES2019' or 'ESNEXT'. */
-    "module": "commonjs",
-    /* Specify module code generation: 'none', 'commonjs', 'amd', 'system', 'umd', 'es2015', or 'ESNext'. */
-    "outDir": "./build",
-    /* Redirect output structure to the directory. */
-    "rootDir": "./src",
-    /* Specify the root directory of input files. Use to control the output directory structure with --outDir. */
-    "strict": true,
-    /* Enable all strict type-checking options. */
-    "noImplicitAny": true,
-    /* Raise error on expressions and declarations with an implied 'any' type. */
-    "esModuleInterop": true,
-    /* Enables emit interoperability between CommonJS and ES Modules via creation of namespace objects for all imports. Implies 'allowSyntheticDefaultImports'. */
-  },
-  "exclude": [
-    "node_modules",
-    "tests"
-  ]
-}
